implementing delete

+++ src/main/java/com/adelacebal/library/web/controller/BookController.java
@@ -51,7 +51,9 @@ public class BookController {
     @RequestMapping(value = "book/delete/{bookId}", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public List<Book> deleteBook(@PathVariable Long bookId) {
         Book book = bookService.findById(bookId);
-        bookService.delete(book);
+        if (book.getBooksIssued() <= 0) {
+            bookService.delete(book);
+        }
 
         List<Book> books = bookService.findAll();
         return books;
